Modularize api stuff - Not tested (need to test on my mac)
